fix: CORUI-6264: One component per file Menu (#434)
* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Rename file

* Rename temp

* Restore original

* Separate components

* Move Menu subcomponents

* Update snapshot with correct on click function
