+++ src/ComboboxInput/ComboboxInput.test.js
@@ -1,20 +1,20 @@
 import { ComboboxInput } from './ComboboxInput';
+import Menu from '../Menu/Menu';
 import { mount } from 'enzyme';
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { Menu, MenuItem, MenuList } from '../Menu/Menu';
 
 describe('<ComboboxInput />', () => {
     const defaultMenu = (
         <Menu>
-            <MenuList>
-                <MenuItem url='/'>Pear</MenuItem>
-                <MenuItem url='/'>Strawberry</MenuItem>
-                <MenuItem url='/'>Raspberry</MenuItem>
-                <MenuItem isLink url='/'>
+            <Menu.List>
+                <Menu.Item url='/'>Pear</Menu.Item>
+                <Menu.Item url='/'>Strawberry</Menu.Item>
+                <Menu.Item url='/'>Raspberry</Menu.Item>
+                <Menu.Item isLink url='/'>
                     + New Item
-                </MenuItem>
-            </MenuList>
+                </Menu.Item>
+            </Menu.List>
         </Menu>
     );
