+++ src/Dropdown/Dropdown.test.js
@@ -1,20 +1,20 @@
 import { Button } from '../';
 import { Dropdown } from './Dropdown';
+import Menu from '../Menu/Menu';
 import { mount } from 'enzyme';
 import { Popover } from '../Popover/Popover';
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { Menu, MenuItem, MenuList } from '../Menu/Menu';
 
 describe('<Dropdown />', () => {
     const defaultMenu = (
         <Menu>
-            <MenuList>
-                <MenuItem url='/'>Option 1</MenuItem>
-                <MenuItem url='/'>Option 2</MenuItem>
-                <MenuItem url='/'>Option 3</MenuItem>
-                <MenuItem url='/'>Option 4</MenuItem>
-            </MenuList>
+            <Menu.List>
+                <Menu.Item url='/'>Option 1</Menu.Item>
+                <Menu.Item url='/'>Option 2</Menu.Item>
+                <Menu.Item url='/'>Option 3</Menu.Item>
+                <Menu.Item url='/'>Option 4</Menu.Item>
+            </Menu.List>
         </Menu>
     );
